move and prune redundant tests
