+++ intake/tests/views/test_county_application_view.py
@@ -1,11 +1,13 @@
+import random
 from unittest.mock import patch
+from django.core.urlresolvers import reverse
+from formation import fields
+from markupsafe import escape
 from intake.tests.views.test_applicant_form_view_base \
     import ApplicantFormViewBaseTestCase
 from intake.views.county_application_view import WARNING_FLASH_MESSAGE
-from django.core.urlresolvers import reverse
-from intake.tests import mock
-from formation import fields
-from markupsafe import escape
+from intake.tests import mock, factories
+from intake import models, constants
 
 
 class TestCountyApplicationNoWarningsView(ApplicantFormViewBaseTestCase):
@@ -26,6 +28,10 @@ class TestCountyApplicationNoWarningsView(ApplicantFormViewBaseTestCase):
         self.set_form_session_data(counties=['alameda', 'contracosta'])
         response = self.client.get(reverse(self.view_name))
         self.assertEqual(response.status_code, 200)
+        self.assertIn('Apply to Clear My Record',
+                      response.content.decode('utf-8'))
+        self.assertNotContains(response, "This field is required.")
+        self.assertNotContains(response, "warninglist")
         form = response.context_data['form']
         self.assertFalse(form.is_bound())
         for key in form.get_field_keys():
@@ -79,11 +85,52 @@ class TestCountyApplicationNoWarningsView(ApplicantFormViewBaseTestCase):
         self.assertContains(
             response, escape(fields.FirstName.is_required_error_message))
 
+    def test_bad_contact_data_shows_expected_errors(self):
+        self.be_anonymous()
+        contracosta = constants.Counties.CONTRA_COSTA
+        answers = mock.fake.contra_costa_county_form_answers()
+        result = self.client.fill_form(
+            reverse('intake-apply'),
+            counties=[contracosta],
+            confirm_county_selection='yes')
+
+        # check for the preferred contact methods validator
+        bad_data = answers.copy()
+        bad_data['contact_preferences'] = ['prefers_email', 'prefers_sms']
+        bad_data['email'] = ''
+        bad_data['phone_number'] = ''
+        result = self.client.fill_form(
+            reverse('intake-county_application'),
+            **bad_data)
+        self.assertTrue(result.context_data['form'].email.errors)
+        self.assertTrue(result.context_data['form'].phone_number.errors)
+
     def test_shows_counties_where_applying(self):
         self.set_form_session_data(counties=['sanfrancisco'])
         response = self.client.get(reverse(self.view_name))
         self.assertContains(
             response, "You are applying for help in San Francisco County.")
+
+    def test_can_go_back_and_reset_counties(self):
+        self.be_anonymous()
+        county_slugs = [slug for slug, text in constants.COUNTY_CHOICES]
+        first_choices = random.sample(county_slugs, 2)
+        second_choices = [random.choice(county_slugs)]
+        self.client.fill_form(
+            reverse('intake-apply'),
+            counties=first_choices,
+            confirm_county_selection='yes',
+            follow=True)
+        county_setting = self.client.session['form_in_progress']['counties']
+        self.assertEqual(county_setting, first_choices)
+
+        self.client.fill_form(
+            reverse('intake-apply'),
+            counties=second_choices,
+            confirm_county_selection='yes',
+            follow=True)
+        county_setting = self.client.session['form_in_progress']['counties']
+        self.assertEqual(county_setting, second_choices)
 
     @patch('intake.services.events_service.log_form_page_complete')
     def test_logs_page_complete_event(self, event_log):
@@ -113,8 +160,15 @@ class TestCountyApplicationNoWarningsView(ApplicantFormViewBaseTestCase):
 class TestCountyApplicationView(TestCountyApplicationNoWarningsView):
     view_name = 'intake-county_application'
 
-    def test_validation_warnings(self):
-        self.set_form_session_data(counties=['sanfrancisco'])
+    @patch(
+        'intake.services.submissions.send_confirmation_notifications')
+    @patch(
+        'intake.views.session_view_base.notifications'
+        '.slack_new_submission.send')
+    def test_validation_warnings(self, slack, send_confirmation):
+        applicant = factories.ApplicantFactory.create()
+        self.set_form_session_data(
+            counties=['sanfrancisco'], applicant_id=applicant.id)
         response = self.client.fill_form(
             reverse(self.view_name),
             **mock.fake.sf_pubdef_answers(ssn=''))
@@ -126,3 +180,8 @@ class TestCountyApplicationView(TestCountyApplicationNoWarningsView):
             response,
             escape(
                 fields.SocialSecurityNumberField.is_recommended_error_message))
+        slack.assert_not_called()
+        send_confirmation.assert_not_called()
+        submitted_event_count = applicant.events.filter(
+            name=models.ApplicationEvent.APPLICATION_SUBMITTED).count()
+        self.assertEqual(0, submitted_event_count)
